pi/requirements.txt
smbus2
pyserial
influxdb-client
python-dotenv